Split OCP data generator to its own class.
